requirements.txt
Flask
gunicorn
Jinja2
psycopg2
SQLAlchemy
urllib3
psycopg2-binary
ujson
redis
newrelic
uuid
flask-bootstrap
boto3
kafka-python
heroku-kafka
pika
cognitive_face
pillow
pusher